ラベル生成関係修正

+++ ProductDataBase/Product/ProductBarcodeSetting.cs
@@ -15,12 +15,12 @@ namespace ProductDatabase.Product {
     public class CBarcodeProPageSettings {
         public int NumLabelsX { get; set; } = 10;
         public int NumLabelsY { get; set; } = 31;
-        public double SizeX { get; set; } = 15;
-        public double SizeY { get; set; } = 5;
-        public double OffsetX { get; set; } = 12;
-        public double OffsetY { get; set; } = 11;
-        public double IntervalX { get; set; } = 4;
-        public double IntervalY { get; set; } = 4;
+        public decimal SizeX { get; set; } = 15;
+        public decimal SizeY { get; set; } = 5;
+        public decimal OffsetX { get; set; } = 12;
+        public decimal OffsetY { get; set; } = 11;
+        public decimal IntervalX { get; set; } = 4;
+        public decimal IntervalY { get; set; } = 4;
         public string HeaderString { get; set; } = "%D 製番 %M 注番 %O %T 台数 %N 担当者 %U";
         public string FooterString { get; set; } = String.Empty;
         public Point HeaderPos { get; set; } = Point.Empty;
@@ -44,13 +44,13 @@ namespace ProductDatabase.Product {
     }
 
     public class CBarcodeProLabelSettings {
-        public double BarcodeHeight { get; set; } = 1;
-        public double BarcodePosX { get; set; } = 4;
-        public double BarcodePosY { get; set; } = 1;
-        public double BarcodeMagnitude { get; set; } = 1.0;
+        public decimal BarcodeHeight { get; set; } = 1;
+        public decimal BarcodePosX { get; set; } = 4;
+        public decimal BarcodePosY { get; set; } = 1;
+        public decimal BarcodeMagnitude { get; set; } = 1.0M;
         public string Format { get; set; } = "%T%Y%MM%S%R";
-        public double StringPosX { get; set; } = 4;
-        public double StringPosY { get; set; } = 1;
+        public decimal StringPosX { get; set; } = 4;
+        public decimal StringPosY { get; set; } = 1;
         public bool AlignStringCenter { get; set; } = true;
         public bool AlignBarcodeCenter { get; set; } = true;
         public int NumLabels { get; set; } = 1;
